requirements.txt
Flask
python-dotenv
comtypes
geocoder
beautifulsoup4
requests
pywhatkit
pyttsx3
gunicorn
